BTN writing function updated. It was too smart.
If the nprs value is negative, the BTN package should write that negative value no matter what was set for timprs. Before there was some logic included that if the user would assign any value to timprs, the nprs was not honoured. For me this was like hidden magic, that I was not expecting from reading the manual.

An alternative solution would be to state in the manual that timprs should be None.

Same goes for the observation item.